# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: hybrd API
  version: 0.1.0
  description: Local-first training control plane with versioned aggregate sync. Better Auth endpoints
    live under /api/auth and use the provider's contract.
servers:
  - url: http://localhost:3000
    description: Local development; override in the generated client for other environments.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Better Auth session token. Store it in the iOS Keychain.
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: {}
            requestId:
              type: string
              format: uuid
          required:
            - code
            - message
            - requestId
      required:
        - error
    Bootstrap:
      type: object
      properties:
        athlete:
          $ref: "#/components/schemas/Athlete"
        device:
          type: object
          properties:
            registered:
              type: boolean
          required:
            - registered
        capabilities:
          type: object
          properties:
            remoteDecisions:
              type: boolean
            remoteCoach:
              type: boolean
            billing:
              type: boolean
            push:
              type: boolean
          required:
            - remoteDecisions
            - remoteCoach
            - billing
            - push
        sync:
          type: object
          properties:
            available:
              type: boolean
            latestSequence:
              $ref: "#/components/schemas/SyncCursor"
          required:
            - available
            - latestSequence
        policy:
          type:
            - object
            - "null"
          properties:
            version:
              type: integer
            checksum:
              type: string
          required:
            - version
            - checksum
        entitlements:
          type: array
          items:
            $ref: "#/components/schemas/Entitlement"
      required:
        - athlete
        - device
        - capabilities
        - sync
        - policy
        - entitlements
    Athlete:
      type: object
      properties:
        id:
          type: string
          format: uuid
        timezone:
          type: string
        locale:
          type: string
        distanceUnit:
          type: string
          enum:
            - km
            - mi
        loadUnit:
          type: string
          enum:
            - kg
            - lb
        weekStartsOn:
          type: integer
          minimum: 1
          maximum: 7
        cloudAiConsent:
          type: boolean
          default: false
        trainingDayBoundary:
          type:
            - string
            - "null"
          default: null
        revision:
          $ref: "#/components/schemas/SyncCursor"
      required:
        - id
        - timezone
        - locale
        - distanceUnit
        - loadUnit
        - weekStartsOn
        - revision
    SyncCursor:
      type: string
      maxLength: 19
      pattern: ^(0|[1-9][0-9]*)$
      example: "0"
    Entitlement:
      type: object
      properties:
        key:
          type: string
        status:
          type: string
          enum:
            - active
            - grace
            - expired
            - revoked
        validUntil:
          type:
            - string
            - "null"
          format: date-time
      required:
        - key
        - status
        - validUntil
    TrainingPolicy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          exclusiveMinimum: 0
        schemaVersion:
          type: integer
          exclusiveMinimum: 0
        checksum:
          type: string
        config:
          $ref: "#/components/schemas/PolicyConfig"
        minimumAppVersion:
          type:
            - string
            - "null"
      required:
        - id
        - version
        - schemaVersion
        - checksum
        - config
        - minimumAppVersion
...
    RegisteredDevice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        registered:
          type: boolean
          enum:
            - true
      required:
        - id
        - registered
    DeviceRegistration:
      type: object
      properties:
        appVersion:
          type: string
          minLength: 1
          maxLength: 64
        osVersion:
          type: string
          minLength: 1
          maxLength: 64
        pushToken:
          type:
            - string
            - "null"
          minLength: 32
          maxLength: 512
          pattern: ^[a-fA-F0-9]+$
        pushEnvironment:
          type: string
          enum:
            - sandbox
            - production
          default: sandbox
        pushEnabled:
          type: boolean
          default: false
      required:
        - appVersion
      additionalProperties: false
...
    ExerciseCatalog:
      type: object
      properties:
        version:
          type: integer
        exercises:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              slug:
                type: string
              name:
                type: string
              movementPattern:
                type:
                  - string
                  - "null"
              unilateral:
                type: boolean
              active:
                type: boolean
              metadata:
                type: object
                additionalProperties: {}
            required:
              - id
              - slug
              - name
              - movementPattern
              - unilateral
              - active
              - metadata
        equipment:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              slug:
                type: string
              name:
                type: string
            required:
              - id
              - slug
              - name
        muscleGroups:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              slug:
                type: string
              name:
                type: string
            required:
              - id
              - slug
              - name
        aliases:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              exerciseId:
                type: string
                format: uuid
              source:
                type: string
              alias:
                type: string
            required:
              - id
              - exerciseId
              - source
              - alias
        exerciseMuscles:
          type: array
          items:
            type: object
            properties:
              exerciseId:
                type: string
                format: uuid
              muscleGroupId:
                type: string
                format: uuid
              role:
                type: string
                enum:
                  - primary
                  - secondary
            required:
              - exerciseId
              - muscleGroupId
              - role
        exerciseEquipment:
          type: array
          items:
            type: object
            properties:
              exerciseId:
                type: string
                format: uuid
              equipmentId:
                type: string
                format: uuid
              required:
                type: boolean
            required:
              - exerciseId
              - equipmentId
              - required
      required:
        - version
        - exercises
        - equipment
        - muscleGroups
        - aliases
        - exerciseMuscles
        - exerciseEquipment
...
paths:
...
  /v1/bootstrap:
    get:
      operationId: getBootstrap
      tags:
        - Athlete
      security: &a3
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: false
          name: deviceId
          in: query
      responses:
        "200":
          description: Account state and sync capabilities.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bootstrap"
        "400":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/config/training-policy:
    get:
      operationId: getTrainingPolicy
      tags:
        - Configuration
      security: *a3
      parameters:
        - schema:
            type: string
          required: false
          name: if-none-match
          in: header
      responses:
        "200":
          description: Latest published policy. Cache only if schemaVersion is supported.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingPolicy"
        "304":
          description: Cached policy is current.
        "400":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/devices/{deviceId}:
    put:
      operationId: registerDevice
      tags:
        - Devices
      security: *a3
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: deviceId
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRegistration"
      responses:
        "200":
          description: Idempotent registration of the authenticated athlete's installation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisteredDevice"
        "400":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: revokeDevice
      tags:
        - Devices
      security: *a3
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: deviceId
          in: path
      responses:
        "204":
          description: Installation revoked. Repeated requests have the same effect.
        "400":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/billing/entitlements:
    get:
      operationId: getEntitlements
      tags:
        - Billing
      security: *a3
      responses:
        "200":
          description: Canonical entitlement records; an empty array grants no access.
          content:
            application/json:
              schema:
                type: object
                properties:
                  entitlements:
                    type: array
                    items:
                      $ref: "#/components/schemas/Entitlement"
                required:
                  - entitlements
        "400":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /v1/account:
    delete:
      operationId: deleteAccount
      tags:
        - Account
      security: *a3
      summary: Permanently delete the account and its data; requires a session created in the last 10
        minutes.
      responses:
        "204":
          description: Account, sessions, training data and provider records purged.
        "400":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/catalog:
    get:
      operationId: getCatalog
      tags:
        - Catalog
      security: *a3
      responses:
        "200":
          description: Versioned exercise, muscle and equipment reference catalog.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseCatalog"
        "400":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/account/export:
    get:
      operationId: exportAccount
      tags:
        - Account
      security: *a3
      responses:
        "200":
          description: Canonical training and coaching data export. Excludes credentials and operational logs.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
        "400":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Request failed. Use the stable error code for client behavior.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...